requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
PyGithub
python-gitlab
bandit
python-dotenv
Jinja2